requirements.txt
numpy==1.15.4
opencv-python
opencv-contrib-python
Pillow